Fix systemd cgroup after memory type changed

LGTMs: @crosbymichael @cyphar
Closes #1573

+++ libcontainer/cgroups/systemd/apply_systemd.go
@@ -260,7 +260,7 @@ func (m *Manager) Apply(pid int) error {
 
 	if c.Resources.Memory != 0 {
 		properties = append(properties,
-			newProp("MemoryLimit", c.Resources.Memory))
+			newProp("MemoryLimit", uint64(c.Resources.Memory)))
 	}
 
 	if c.Resources.CpuShares != 0 {
